Create html: countDown
This code creates a simple countdown timer that starts when the "Start" button is clicked. It displays a countdown from 10 seconds and updates the time displayed every second. Once the countdown reaches zero, the timer stops. The functionality is implemented using the setInterval() method to update the time and clearInterval() to stop the countdown once it reaches 0.